+++ programs/monaco_protocol/src/lib.rs
@@ -5,9 +5,7 @@ use crate::error::CoreError;
 use crate::instructions::market::verify_market_authority;
 use crate::instructions::transfer;
 use crate::instructions::verify_operator_authority;
-use crate::state::market_account::{
-    Market, MarketMatchingPool, MarketOrderBehaviour, MarketOutcome,
-};
+use crate::state::market_account::{Market, MarketOrderBehaviour};
 use crate::state::market_position_account::MarketPosition;
 use crate::state::market_type::verify_market_type;
 use crate::state::operator_account::AuthorisedOperators;
