+++ programs/monaco_protocol/src/state/mod.rs
@@ -1,4 +1,6 @@
 pub(crate) mod market_account;
+pub(crate) mod market_matching_pool_account;
+pub(crate) mod market_outcome_account;
 pub(crate) mod market_position_account;
 pub(crate) mod market_type;
 pub(crate) mod operator_account;
